requirements.txt
onnxruntime-qnn==1.18.0
flask>=3
numpy
pillow
pyttsx3
soundfile
tokenizers>=0.15
scipy>=1.11 
